Added forward taco tuning UI elements and modified layout a bit.
