+++ utils/modelutils.py
@@ -10,8 +10,14 @@ def check_model_paths(encoder_path: Path, synthesizer_path: Path, vocoder_path: 
         return
 
     # If none of the paths exist, remind the user to download models if needed
+    model_files_missing()
+    quit(-1)
+
+def model_files_missing(type=None):
     print("********************************************************************************")
-    print("Error: Model files not found. Follow these instructions to get and install the models:")
+    if type is None:
+        print("Error: Model files not found. Follow these instructions to get and install the models:")
+    else:
+        print("Error: %s Model files not found. Follow these instructions to get and install the models:" % type)
     print("https://github.com/CorentinJ/Real-Time-Voice-Cloning/wiki/Pretrained-models")
     print("********************************************************************************\n")
-    quit(-1)
